Change param to key in GeneticAlgorithm.__set_params()
'param' is a confusing name. The abbreviation uses fewer charcters
than 'parameter' would. However, it sacrifices redability for the
sake of brevity. 'param' requires the reader to perform an extra step
of translation each time it is read. Using another name fascilitates
reader comprehension at a time when they should be learning the method's
 underlying logic.

This method employs dictionary comprehension to set the values of the
attribute GeneticAlgorithm.settings. 'key' is a cleaner alternative,
for this name references what is being handled within
GeneticAlgorithm.__set__params().